% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpeff_fun.R
\name{snpeffr}
\alias{snpeffr}
\title{Pull in, filter, and parse vcf data and annotations from snpeff}
\usage{
snpeffr(
  vcf_path,
  positions = list(fks1_hs1 = 221637:221663, fks1_hs2 = 223782:223805, fks1_hs3 =
    221805:221807),
  genes = c("CAB11_002014"),
  exclude_effects = "synonymous_variant"
)
}
\arguments{
\item{vcf_path}{the path to the vcf file output from snpeff (read in using data.table::fread and R.utils if zipped)}

\item{positions}{a named list of positions to filter to, each element can either
be a sequence, i.e. 1000:2000 or a vector i.e. c(1000:2000, 300).}

\item{genes}{the name of the genes of interest to filter to}

\item{exclude_effects}{a string formatted as a regular expression,
effects to ignore (as categorized by snpeff, see \href{https://pcingola.github.io/SnpEff/se_inputoutput/}{here}),}
}
\value{
a data.frame with n rows and 7 columns:
\describe{
\item{\code{sample_id}}{ID of the sample, the names of the annotation columns from snpeff}
\item{\code{snpeff_gene_name}}{name of the gene with the mutation}
\item{\code{region}}{name of the region of interest, i.e. from the named list of positions,
defaults to fks1_hs1, fks1_hs2, and fks1_hs3 (fks1, hotspots 1, 2, & 3)}
\item{\code{position}}{the nucleotide position of the mutation}
\item{\code{mutation}}{the protein level mutation, annotated using (HGVS notation)(http://varnomen.hgvs.org/bg-material/simple/)}
\item{\code{ref_sequence}}{the nucleotide sequence for the reference}
\item{\code{sample_sequence}}{the nucleotide sequence for the sample}
}
}
\description{
This function parses a vcf file generated by snpeff and filters to specific
genes, positions, and effects of interest. See the \href{https://pcingola.github.io/SnpEff}{snpeff}
documentation for more information on
how to interpret these inputs and outputs.
}
